requirements.txt
django==5.1.5
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-telegram-bot==21.10
requests==2.32.3